version 0.2
Somehow the wrong hotkeys got uploaded the last time. Had to redo a lot of the work. Oh well...
This is compliant with TB 11, version 33 now. So a lot of the Mac-exclusive shortcuts have been purged.
I added some icons.  A galaxy to signify unused hotkeys. A comet to show that something is not quite right.
Others to seperate categories visually from one another, should a folder have multiple.